+++ examples/permissioned-token-freezing/assets/clawback-escrow.py
@@ -79,4 +79,4 @@ if __name__ == "__main__":
     if(len(sys.argv) > 1):
         params = parse_params(sys.argv[1], params)
 
-    print(compileTeal(clawback_escrow(params["ASSET_ID"], params["APP_ID"]), Mode.Signature))
+    print(compileTeal(clawback_escrow(params["ASSET_ID"], params["APP_ID"]), Mode.Signature, version = 4))